requirements.txt
ConfigArgParse
websocket-client
requests
paho-mqtt
